Práctica finalizada

+++ include/registers.hpp
@@ -5,17 +5,17 @@
 
 class RegisterMemory {
   private:
-    std::vector<int> regist;
+    std::vector<long long> regist;
   public:
     RegisterMemory();
     RegisterMemory(const int &number);
     ~RegisterMemory();
 
-    int getReg(const int &number);
+    long long getReg(const int &number);
     void setReg(const int &index, int value);
     int numbOfReg(void);
     void writeRegisters();
-    std::vector<int> getAllReg();
+    std::vector<long long> getAllReg();
     void cleanRegisters(void);
     // friend std::ostream& operator<<(std::ostream& os, RegisterMemory& dt);
 };
